+++ src/db_migrate.cpp
@@ -14,7 +14,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.9.19"};
+constexpr auto version = std::string_view{"0.9.20"};
 
 auto print_help() -> void
 {