+++ src/main.cpp
@@ -39,7 +39,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.9.19"};
+constexpr auto version = std::string_view{"0.9.20"};
 
 struct BootstrapConfigResult final
 {